Fix the CUPS driver: don't use the page number to decide whether to initialize the job (use a variable dedicated to that purpose), and don't attempt to throw away a negative amount of data (which would happen in the case of full bleed).
Fix the GIMP plugin: constrain the minimum and maximum sizes of
custom paper sizes based on the printer chosen, and when the printer
changes, change those bounds.  Also set the parameters for the raw
"printer" used to produce the thumbnail preview.

In the core: rather than have fixed pointers to component data for
the dither, color, and driver, make this extensible by storing these
objects as named objects.  Also have them automatically destroyed by
the stp_vars_t being destroyed.  This simplifies the internal interface
and will allow us to extend this scheme to handle more things (such
as a first class interface for channels).

In the Epson driver: remove some now dead code.  Also fix the gray vs.
color command (it should be two bytes, not one).

In the raw pseudo-driver: remove the PageSize option.  We always require
the number of dots to be explicitly set, since there's no real concept
of a "page".

In the weave code: change hard coded array lengths to be named constants.

In the printer code: perform verification on a copy of the vars, rather
than on the real thing.  Also let the driver verify the vars on job
start and end, to match how the print routine works.