Add constraints to PPD files so that Quality setting and color vs. grayscale/black and white can turn off some options.
Fix cups-genppdupdate.  This will not work correctly in 4.3.18; the
PPD files will need to be updated manually one last time.

Add preliminary support for ink limiting (currently disabled).

Send the advance commands for passes with zero advance in the Epson
driver to try to work around some printer problems with printing
very close to the start of the page.

Merge print vars correctly in cases where a user value is not active.