Don't autoload XML files.  Instead, allow components to register XML files for loading, or load them explicitly when needed.
Change the printer code to register its XML file, and the paper
and dither matrix code to load when needed.

In addition to allowing components to intelligently order the
loading of the XML files (and thereby avoid problems with hidden
dependencies and random load order), this speeds up startup because
the dither matrices are large and take on the order of a second
each to load.

Finally, cache the actual dither matrix when loaded, not just the
filename.  This speeds up individual pages.