Add language support for driver interface.
Add a list of language mappings to improve robustness of setlocale.
On some systems, setlocale(LC_ALL, "de") fails, but de_DE works.

Add more default information to PPD files for future use in extracting
setting information.

Change URI encoding to make the driver be the hostname and the
resource be everything else.