Create separate targets for global PPDs, discrete translated PPDs, non-translated PPDs, and global PPDs with numbers "localized".  The latter is for testing purposes, to work around bugs in cupstestppd.
Add -N option to genppd to optionally localize numbers, for testing
purposes.

Treat absence of any -L argument to genppd as an indication to create
globalized PPDs; -L C generates non-translated PPD files.

Add -l option to cups-genppdupdate to specify the language for PPD
files.  For Gutenprint 5.2 and above, this is ignored as globalized
PPD files are corrected, so this is for downgrade purposes (at present).

Do not actually run test-ppds as part of automated testing, due to
problems with cupstestppd.

Add code to filter out spurious errors and warnings in test-ppds.